Add files via upload
Base starting page with Bootstrap and Jquery. I played around with Jquery scroll-event plus background-image changer. Not there yet and we should speak some when about what we think would be the best approach for this git-hub project.
In general, I hope I got it right and now I make a change on the repo :D